+++ classes/prefix_WPgutenberg/blocks/index.js
@@ -2,3 +2,4 @@
  * Import blocks
  */
 import "./vimeo";
+import "./posts";